+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fixes
 
 - `# @jwt` metadata parsedBody matches body (#454)
+- allow comment in last line of script (#454)
 
 ## [6.4.3]